change hash value type to size_t

+++ client/display/painters.c
@@ -59,7 +59,7 @@ draw_world(struct win *win, struct world *w, struct point *view)
 	size_t i;
 	struct point onp, np = onp = nearest_chunk(view);
 	struct rectangle *r = &win->rect;
-	const uint16_t *val;
+	const size_t *val;
 
 	for (; np.x < view->x + r->width; np.x += CHUNK_SIZE) {
 		for (np.y = onp.y; np.y < view->y + r->height; np.y += CHUNK_SIZE) {
